+++ NEWS.md
@@ -1,4 +1,4 @@
-# AMR 1.2.0.9007
+# AMR 1.2.0.9008
 ## <small>Last updated: 17-Jun-2020</small>
 
 ### New
@@ -13,17 +13,23 @@
   tibble(J01CA01 = "S") %>%
     select(penicillins())
   #> Selecting beta-lactams/penicillins: `J01CA01` (ampicillin)
+  
+  # select an antibiotic class manually with `ab_class()`
+  example_isolates %>% 
+    select(ab_class("mycobact"))
+  #> Selecting antimycobacterials: `RIF` (rifampicin)
   ```
 
 ### Changed
 * Fixed a bug where `eucast_rules()` would not work on a tibble when the `tibble` or `dplyr` package was loaded
-* All `*_join_microorganisms()` functions now return the original data class (e.g. tibbles and data.tables)
+* All `*_join_microorganisms()` functions and `bug_drug_combinations()` now return the original data class (e.g. `tibble`s and `data.table`s)
 * Fixed a bug where `as.ab()` would return an error on invalid input values
 * Fixed a bug for using grouped versions of `rsi_df()`, `proportion_df()` and `count_df()`, and fixed a bug where not all different antimicrobial results were added as rows
 * Added function `filter_penicillins()` to filter isolates on a specific result in any column with a name in the antimicrobial 'penicillins' class (more specific: ATC subgroup *Beta-lactam antibacterials, penicillins*)
 * Added official antimicrobial names to all `filter_ab_class()` functions, such as `filter_aminoglycosides()`
 * Added antibiotics code "FOX1" for cefoxitin screening (abbreviation "cfsc") to the `antibiotics` data set
 * Improved auto-determination for columns of types `<mo>` and `<Date>`
+* Fixed a bug in `bug_drug_combinations()` for when only one antibiotic was in the input data
 
 # AMR 1.2.0